backend/pyproject.toml
[project]
name = "adk-hackathon"
version = "0.1.0"
description = ""
authors = [
    {name = "davidtaoweiji",email = "davidtaoweiji@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-adk (>=1.0.0,<2.0.0)",
    "google-auth-oauthlib (>=1.2.2,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
packages = [
    { include = "common" },
    { include = "email_agent" }
]
